move compose-plugin commands under /cmd
