+++ cmd/compose/compose_test.go
