+++ cmd/compose/kill.go
