+++ cmd/compose/remove.go
