+++ cmd/compose/stop.go
