+++ cmd/compose/up_test.go
