+++ docs/yaml/main/generate.go
@@ -27,7 +27,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 
-	"github.com/docker/compose-cli/cli/cmd/compose"
+	"github.com/docker/compose-cli/cmd/compose"
 	. "github.com/docker/compose-cli/docs/yaml"
 )
